fix: progress field f32 to match FLOAT4 column
Migration m20260605_005 uses .float() which creates FLOAT4 (REAL),
but entity used f64 which maps to FLOAT8 (DOUBLE PRECISION).

Error: 'Rust type core::option::Option<f64> (as SQL type FLOAT8) is
not compatible with SQL type FLOAT4'

+++ crates/randimg-core/src/db/entities/task.rs
@@ -48,7 +48,7 @@ pub struct Model {
     pub retry_count: i32,
 
     /// 任务进度百分比 (0.0 ~ 100.0)
-    pub progress: f64,
+    pub progress: f32,
 
     /// 优先级（数值越小，优先级越高）
     pub priority: i32,
